refactor: move event handling to application layer following DDD
BREAKING CHANGE: Refactored email event handling to follow proper DDD layering

Changes:
- Created IEventHandler<TEvent> interface in domain layer
- Moved event handlers from infrastructure to application layer:
  * UserRegisteredEventHandler
  * ApplicationSubmittedEventHandler
  * ApplicationReviewedEventHandler
- Each handler implements IEventHandler interface
- Deleted EmailEventListener from infrastructure layer
- Updated tests to match new structure

This fixes the DDD violation where infrastructure layer
was depending on application layer (EmailService).

Now follows proper dependency flow:
Application → Domain ← Infrastructure ✓

Resolves peer review comment about DDD layer violation.
